requirements.txt
altgraph==0.17
bottle==0.12.18
bottle-websocket==0.2.9
cffi==1.14.0
click==7.1.2
Eel==0.12.4
ffmpeg-python==0.2.0
filetype==1.0.7
future==0.18.2
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pefile==2019.4.18
pycparser==2.20
pydub==0.24.0
PyInstaller==3.6
pywin32-ctypes==0.2.0
six==1.15.0
Werkzeug==1.0.1
whichcraft==0.6.1
